Fix GameServer build warnings + real CheckGameMasterLevel auth bug
A full rebuild (needed for the stdafx.h change) surfaced 35 warnings
that had been silently accumulating in GameServer.vcxproj. Most were
missing _CRT_SECURE_NO_WARNINGS/_WINSOCK_DEPRECATED_NO_WARNINGS
(present in the other 3 server projects, never added here) plus an
LNK4098 CRT mismatch on the CS config. 6 more were real minor issues
(shift-before-widen, narrowing casts, signed/unsigned compare, an
unused local, double->float) - all fixed individually.

The one that mattered: CheckGameMasterLevel (GameMaster.cpp) never
compared the caller's level against the requested one, and returns
bool from an int - so a non-GM player fell through to `return -1`
(truthy) and passed every GM-only check, while a real dynamically-
registered GM at Level 0 got 'false' and failed every check. Fixed to
compare Level >= level and return false (not -1) on no-match. Needs
in-game testing with a real GM account before this is trusted.

Both Release_EX603/Release_EX603CS rebuild clean (0 warnings, 0
errors) and are redeployed to MuServer.

+++ Source_Server/GameServer/GameServer/MemScript.cpp
@@ -5,12 +5,12 @@
 #include "stdafx.h"
 #include "MemScript.h"
 
-// ¡Construction!
+// ï¿½Construction!
 CMemScript::CMemScript() : m_buff(nullptr), m_size(0), m_path{} {
 	SetLastError(0);
 }
 
-// !Destruction¡
+// !Destructionï¿½
 CMemScript::~CMemScript() {
 	delete[] m_buff;
 }
@@ -178,7 +178,7 @@ eTokenResult CMemScript::GetTokenNumber(char ch) // OK
 	}
 	else
 	{
-		this->m_number = atof(this->m_string);
+		this->m_number = (float)atof(this->m_string);
 	}
 
 	this->m_string[count] = '\0';
